requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cu101
torchvision==0.8.2+cu101
matplotlib
scipy
trimesh
scikit-image
tensorboard
numpy
tqdm
torchmeta
pandas
tables
pillow
rtree
scikit-learn
pyrender
mesh_to_sdf
pyglet